test(daemon): split fleet-transport integration test by responsibility
